AppTigres/requirements.txt
Flask
Flask-SQLAlchemy
SQLAlchemy
matplotlib
seaborn
pandas
numpy
google-generativeai
python-dotenv
gunicorn
psycopg2-binary
